+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+- Add Kyori MiniMessage support.
+- The config autoreload on save.
+- The user no longer needs a code to authenticate.
 
 ## [1.0.0] - 2022-09-08